requirements.txt
asttokens==2.4.1
blas==1.0
brotli==1.1.0
brotli-bin==1.1.0
bzip2==1.0.8
ca-certificates==2024.3.11
cached-property==1.5.2
cached_property==1.5.2
certifi==2024.2.2
cffi==1.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
farama-notifications==0.0.4
filelock==3.14.0
fonttools==4.51.0
freetype==2.12.1
fsspec==2024.3.1
gettext==0.22.5
gettext-tools==0.22.5
glib==2.80.0
glib-tools==2.80.0
gst-plugins-base==1.24.1
gstreamer==1.24.1
gymnasium==0.29.1
h5py==3.11.0
hdf5==1.14.3
icu==73.2
importlib-metadata==7.1.0
importlib_metadata==7.1.0
intel-openmp==2023.2.0
ipykernel==6.29.3
ipython==8.22.2
jax-jumpy==1.0.0
jedi==0.19.1
jinja2==3.1.3
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
krb5==1.21.2
lcms2==2.16
lerc==4.0.0
libaec==1.1.3
libasprintf==0.22.5
libasprintf-devel==0.22.5
libblas==3.9.0
libbrotlicommon==1.1.0
libbrotlidec==1.1.0
libbrotlienc==1.1.0
libcblas==3.9.0
libclang13==18.1.3
libcurl==8.7.1
libdeflate==1.20
libexpat==2.6.2
libffi==3.4.2
libgettextpo==0.22.5
libgettextpo-devel==0.22.5
libglib==2.80.0
libhwloc==2.10.0
libiconv==1.17
libintl==0.22.5
libintl-devel==0.22.5
libjpeg-turbo==3.0.0
liblapack==3.9.0
libogg==1.3.4
libpng==1.6.43
libsodium==1.0.18
libsqlite==3.45.3
libssh2==1.11.0
libtiff==4.6.0
libuv==1.48.0
libvorbis==1.3.7
libwebp-base==1.4.0
libxcb==1.15
libxml2==2.12.6
libzlib==1.2.13
m2w64-gcc-libgfortran==5.3.0
m2w64-gcc-libs==5.3.0
m2w64-gcc-libs-core==5.3.0
m2w64-gmp==6.1.0
m2w64-libwinpthread-git==5.0.0.4634.697f757
markupsafe==2.1.5
matplotlib==3.8.4
matplotlib-base==3.8.4
matplotlib-inline==0.1.7
mkl==2023.2.0
mkl-service==2.4.1
mpmath==1.3.0
msys2-conda-epoch==20160418
munkres==1.1.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.12.0
numpy==1.26.4
openjpeg==2.5.2
openssl==3.3.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pcre2==10.43
pickleshare==0.7.5
pillow==10.3.0
pip==24.0
platformdirs==4.2.1
ply==3.11
prompt-toolkit==3.0.42
psutil==5.9.8
pthread-stubs==0.4
pthreads-win32==2.9.1
pure_eval==0.2.2
pycparser==2.22
pygments==2.17.2
pyparsing==3.1.2
pyqt==5.15.9
pyqt5-sip==12.12.2
python==3.12.3
python-dateutil==2.9.0
python-tzdata==2024.1
python_abi==3.12
pytorch==2.2.0
pytz==2024.1
pywin32==306
pyyaml==6.0.1
pyzmq==26.0.2
qt-main==5.15.8
setuptools==69.5.1
sip==6.7.12
six==1.16.0
stack_data==0.6.2
sympy==1.12
tbb==2021.12.0
tk==8.6.13
toml==0.10.2
tomli==2.0.1
tornado==6.4
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024a
ucrt==10.0.22621.0
vc==14.3
vc14_runtime==14.38.33130
vs2015_runtime==14.38.33130
wcwidth==0.2.13
wheel==0.43.0
xorg-libxau==1.0.11
xorg-libxdmcp==1.1.3
xz==5.2.6
yaml==0.2.5
zeromq==4.3.5
zipp==3.17.0
zstd==1.5.5
